backend/requirements.txt
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
sqlparse==0.5.3
